VB-1247 Remove legacy update and create end points from VSiP backend, moved non pageable get visits to legacy controller
